+++ src/recordlinker/models/__init__.py
@@ -1,17 +1,21 @@
-from sqlalchemy import create_engine
-from sqlalchemy import orm
+from .algorithm import Algorithm
+from .algorithm import AlgorithmPass
+from .base import Base
+from .base import get_session
+from .mpi import BlockingKey
+from .mpi import BlockingValue
+from .mpi import Patient
+from .mpi import Person
+from .pii import PIIRecord
 
-from recordlinker.config import settings
-
-
-def get_session() -> orm.Session:
-    """
-    Creates a new session to the MPI database and returns it.
-    """
-    engine = create_engine(settings.db_uri)
-    Base.metadata.create_all(engine)
-    return orm.Session(engine)
-
-
-class Base(orm.DeclarativeBase):
-    pass
+__all__ = [
+    "Base",
+    "get_session",
+    "PIIRecord",
+    "Person",
+    "Patient",
+    "BlockingKey",
+    "BlockingValue",
+    "Algorithm",
+    "AlgorithmPass",
+]
