+++ src/recordlinker/models/mpi.py
@@ -5,7 +5,7 @@ from sqlalchemy import orm
 from sqlalchemy import schema
 from sqlalchemy import types as sqltypes
 
-from . import Base
+from .base import Base
 from .pii import PIIRecord
